Content changes:
This commit comes with some small changes to the "eml is valid" test. It
removes unnecessary stuff from the test, improves the comments and the most
important thing is, it makes the test work properly. The response from the body
is a string and so the xml document generation did not work with the method
".document". This commit changes the method to ".string".